requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
boto3==1.18.39
botocore==1.21.39
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.6
django-autocomplete-light==3.8.2
django-bootstrap5==2.1.2
django-contrib-comments==2.1.0
django-extensions==3.1.3
django-froala-editor==3.2.2
django-heroku==0.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-taggit==1.5.1
gunicorn==20.1.0
idna==3.2
Pillow==8.3.1
python-dateutil==2.8.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
Unidecode==1.3.1
urllib3==1.26.6
Werkzeug==2.0.1
whitenoise==5.3.0
